requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
bidict==0.23.1
billiard==4.2.2
biosppy==2.2.4
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.8
djangorestframework==3.16.1
fastparquet==2024.11.0
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
h5py==3.15.1
heartpy==1.2.7
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.7
mock==5.2.0
neurokit2==0.2.12
numpy==2.2.6
openai==2.7.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyWavelets==1.9.0
redis==7.0.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
